requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.5.1
carla==0.9.13
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==39.0.1
cycler==0.11.0
Cython==3.0.11
cython-bbox==0.1.5
distro==1.9.0
dm-tree==0.1.8
exceptiongroup==1.2.2
filterpy==1.4.5
flatbuffers==25.1.24
fonttools==4.38.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.2
imagesize==1.4.1
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
lap==0.5.12
loguru==0.7.3
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
motmetrics==1.4.0
MouseInfo==0.1.3
mss==7.0.1
networkx==2.6.3
ninja==1.11.1.3
numpy==1.21.6
nvidia-dali-cuda110==1.31.0
oauthlib==3.2.2
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.8.0
openai==1.39.0
opencv-python==4.11.0.86
packaging==24.0
pandas==1.3.5
pathspec==0.11.2
Pillow==9.4.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycocotools
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.1
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.7.1
ujson==5.7.0
ultralytics==8.0.145
urllib3==1.26.14
Werkzeug==2.2.3
xmltodict==0.14.2
git+https://github.com/ifzhang/ByteTracker.git#egg=yolox
zipp==3.15.0
